SurveyWrapper: partial revert ef379c22a833ad8f0e71b42740daabc19dcb299d
one of my older fixes in this files was likely wrong.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/external/scipio-ide/issues/5

+++ applications/content/src/org/ofbiz/content/survey/SurveyWrapper.java
@@ -677,9 +677,7 @@ public class SurveyWrapper {
         switch (type) {
             case 1:
                 if (result[0] > 0) {
-                    // SCIPIO: TODO: REVIEW: only one of these is right...
-                    //result[2] = result[1] / ((long) result[0]);
-                    result[2] = ((long) result[1]) / ((long) result[0]);
+                    result[2] = result[1] / ((long) result[0]);
                 }
                 break;
             case 2:
@@ -689,9 +687,7 @@ public class SurveyWrapper {
                 break;
             case 3:
                 if (result[0] > 0) {
-                    // SCIPIO: TODO: REVIEW: only one of these is right...
-                    //result[2] = result[1] / (long) result[0];
-                    result[2] = result[1] / result[0];
+                    result[2] = result[1] / (long) result[0];
                 }
                 break;
         }
